fix: correct conversion int64 float64 (#278)
* fix: correct conversion int64 float64


* update manifests


* Allow use of dangerous types in CRD


* Update api/v1alpha1/paasconfig_types.go


* make sure to update manifests


* Ensure ratio is between 0 and 1


---------

+++ api/v1alpha1/paasconfig_types.go
@@ -228,7 +228,10 @@ type ConfigQuotaSettings struct {
 
 	// The ratio of the requested quota which will be applied to the total quota
 	// +kubebuilder:validation:Optional
-	Ratio int64 `json:"ratio"`
+	// +kubebuilder:validation:Format:=float
+	// +kubebuilder:validation:Minimum:=0.0
+	// +kubebuilder:validation:Maximum:=1.0
+	Ratio float64 `json:"ratio"`
 
 	// The default quota which the enabled capability gets
 	// +kubebuilder:validation:Required
